refactoring + Readme
