+++ README.md
@@ -1,20 +1,68 @@
-# Introduction 
-TODO: Give a short introduction of your project. Let this section explain the objectives or the motivation behind this project. 
+# Introduction
 
-# Getting Started
-TODO: Guide users through getting your code up and running on their own system. In this section you can talk about:
-1.	Installation process
-2.	Software dependencies
-3.	Latest releases
-4.	API references
+Authentication/Authorization API, this API is responsible any service related to the authentication and authorization of
+the
+user like authenticating the users, issuing the access/refresh tokens, api keys, validating the permissions for the
+users.
 
-# Build and Test
-TODO: Describe and show how to build your code and run the tests. 
+## System requirements
 
-# Contribute
-TODO: Explain how other users and developers can contribute to make your code better. 
+| Software | Version |
+|----------|---------|
+| Java     | 17      |
+| Maven    | > 3     |
+| Spring   | 2.7.4   |
 
-If you want to learn more about creating good readme files then refer the following [guidelines](https://docs.microsoft.com/en-us/azure/devops/repos/git/create-a-readme?view=azure-devops). You can also seek inspiration from the below readme files:
-- [ASP.NET Core](https://github.com/aspnet/Home)
-- [Visual Studio Code](https://github.com/Microsoft/vscode)
-- [Chakra Core](https://github.com/Microsoft/ChakraCore)
+## Development IDE tool preference
+
+Eclipse (https://www.eclipse.org/downloads/)
+
+Intellij (https://www.jetbrains.com/idea/download)
+
+## Dependencies
+
+Other third-party dependencies used in the project please browse the Maven pom.xml file for details of libraries and
+versions used.
+
+## Cloning the project
+
+Clone the project from source controller
+
+```aidl
+$ git clone https://TDEI-UW@dev.azure.com/TDEI-UW/TDEI/_git/gateway
+```
+
+## Building the project
+
+### IDE
+
+Import the project in your preferred IDE by selecting POM.xml.
+
+### Cli
+
+*Note: Navigate to the cloned repository directory before proceeding for below steps
+
+### 1. Building the server
+
+Use Maven command to build the server. Below mvn command will
+
+1. Cleans the project and removes all files generated by the previous build
+2. Builds the maven project and generates the jar package to target directory
+
+```
+$ mvn clean install
+```
+
+### 2. Running the server
+
+```
+$ cd target
+$ java -jar -Dspring.profiles.active=dev gateway-0.0.1.jar
+```
+
+### 3. Browse API documentation
+
+Navigate to the below link for API documentation and API playground
+
+http://localhost:8080/swagger-ui/index.html
+
