+++ src/main/java/com/tdei/auth/core/config/ApiDocConfig.java
@@ -1,4 +1,4 @@
-package com.tdei.auth.config;
+package com.tdei.auth.core.config;
 
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Contact;
