+++ src/main/java/com/tdei/auth/core/config/exception/handler/exceptions/InvalidCredentialsException.java
@@ -1,4 +1,4 @@
-package com.tdei.auth.config.exception.handler.exceptions;
+package com.tdei.auth.core.config.exception.handler.exceptions;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
